refactor: Dealer와 Player가 Participant를 상속하도록 변경
